ansi: implemented Pieco's (Pablo Draw) "24-bit ANSI" sequence
This is an ugly hack, but then again, it's a lame sequence. If this
sequence is detected in the piece, the render engine will fall back
to the EGA palette and then allow for up to 240 custom colours. So
you will not be able to use this in conjunction with the 256 color
ANSI sequences, nor with pieces that have a very rich palette.

Anyway, the sequence should never have come into existance, so this
will do for now.